Moved expfit to utility submodule
